requirements.txt
discord.py==1.7.3
schedule==1.1.0
Pillow==8.3.1